requirements.txt
# Dependencias del proyecto clasificador de autos y bicicletas

# Procesamiento de imágenes
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Análisis de datos
numpy>=1.24.0
pandas>=2.0.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Interfaz gráfica
tkinter  # Incluido con Python

# Utilidades
pathlib  # Incluido con Python
pickle   # Incluido con Python

# Desarrollo y testing (opcional)
jupyter>=1.0.0
pytest>=7.0.0

# Para notebooks (opcional)
ipykernel>=6.0.0
keras>=2.13.0
pyQt5>=5.15.0
joblib>=1.3.0